check CurvatureEncoder logic

+++ Strange_Lovers/DetectorPart/CurvatureEncoder.h
@@ -11,8 +11,8 @@ typedef struct{
 }CurvatureEncoder;
 
 
-int CurvatureEncoder_get_curvature(CurvatureEncoder *this_CurvatureEncoder);
-int CurvatureEncoder_calc_curvature(CurvatureEncoder *this_CurvatureEncoder);
+float CurvatureEncoder_get_curvature(CurvatureEncoder *this_CurvatureEncoder);
+float CurvatureEncoder_calc_curvature(CurvatureEncoder *this_CurvatureEncoder);
 void CurvatureEncoder_init(CurvatureEncoder *this_CurvatureEncoder);
 
 #endif
