+++ Strange_Lovers/DetectorPart/DirectionEncoder.h
@@ -16,8 +16,8 @@ typedef struct{
 }DirectionEncoder;
 
 
-int DirectionEncoder_get_direction(DirectionEncoder *this_DirectionEncoder);
-int DirectionEncoder_calc_speed(DirectionEncoder *this_DirectionEncoder);
+float DirectionEncoder_get_direction(DirectionEncoder *this_DirectionEncoder);
+float DirectionEncoder_calc_speed(DirectionEncoder *this_DirectionEncoder);
 void DirectionEncoder_init(DirectionEncoder *this_DirectionEncoder);
 
 #endif
